+++ site/dat/docs/Changelog.md
@@ -1,9 +1,24 @@
 # Changelog
+## 1.6.7 (next)
+ - add worker id to cloud log file names
+ - add `cloud` and `local` aliases 
+ - fix delay trouble in provisioning
+ - fix Gatling keepalive type cast bug
+ - add `detach` mode to Cloud Provisioning
+ - fix grinder in the cloud with requests scenario
+ - add `xml-jtl-flags` option for tuning of logging verbosity
+ - add `waiting for data...` into titles of console blocks
+ - fix SSL error handling for cloud provisioning
+ - add Chrome profiler [service](ChromeProfiler.md) and reporter
+ - add webdriver.log for requests-based Selenium tests
 
-## 1.6.6 (next)
+## 1.6.6 <sup>08 aug 2016</sup>
  - optimize aggregator by removing excessive calls to `BetterDict.get()`
  - use JMeter Plugins Manager for installation of plugins
  - build installer for Windows to simplify installation process
+ - add Proxy2JMX service to convert tests from Selenium to JMeter format
+ - fix occasional crash from inconsistent API result for cloud test KPIs
+ - prefer user-supplied `download-link` over default download link for JMeter
 
 ## 1.6.5 <sup>12 jul 2016</sup>
  - fix master_id choosing for cloud reporting